+++ src/Plugin/PluginManager.php
@@ -224,5 +224,4 @@ final class PluginManager implements Pluggable {
             }
         });
     }
-
 }