+++ test/Stub/CustomPluginInterface.php
@@ -11,5 +11,4 @@ use Cspray\Labrador\Plugin\Plugin;
 interface CustomPluginInterface extends Plugin {
 
     public function myMethod();
-
 }