+++ test/Stub/PluginDependsNotPluginStub.php
@@ -3,7 +3,6 @@
 
 namespace Cspray\Labrador\Test\Stub;
 
-
 use Cspray\Labrador\Engine;
 use Cspray\Labrador\Plugin\PluginDependentPlugin;
 
@@ -12,5 +11,4 @@ class PluginDependsNotPluginStub implements PluginDependentPlugin {
     public static function dependsOn(): array {
         return [Engine::class];
     }
-
 }